refactor: extract benchmark ledger maintenance CLI commands
